Remove external surfaceTexture related code
