+++ letsrobotandroid-api/src/main/java/tv/letsrobot/android/api/components/camera/SurfaceTextureCameraBaseComponent.kt
@@ -1,0 +1,38 @@
+package tv.letsrobot.android.api.components.camera
+
+import android.content.Context
+import tv.letsrobot.android.api.models.CameraSettings
+import tv.letsrobot.android.api.utils.EglCore
+import tv.letsrobot.android.api.utils.SurfaceTextureUtils
+import javax.microedition.khronos.egl.EGLSurface
+
+/**
+ * Created by Brendon on 10/6/2018.
+ */
+abstract class SurfaceTextureCameraBaseComponent(context: Context, settings: CameraSettings) : CameraBaseComponent(context, settings){
+    protected abstract fun setupCamera()
+    protected abstract fun releaseCamera()
+    protected var eglCore: EglCore? = null
+    protected var eglSurface: EGLSurface? = null
+    protected var mStManager : SurfaceTextureUtils.SurfaceTextureManager
+    init {
+        eglCore = EglCore()
+        eglSurface = eglCore?.createOffscreenSurface(640, 480)
+        eglCore?.makeCurrent(eglSurface)
+        mStManager = SurfaceTextureUtils.SurfaceTextureManager()
+    }
+
+    override fun enableInternal() {
+        super.enableInternal()
+        setupCamera()
+    }
+
+    override fun disableInternal() {
+        super.disableInternal()
+        releaseCamera()
+    }
+
+    open fun init(){
+
+    }
+}
