+++ letsrobotandroid-api/src/main/java/tv/letsrobot/android/api/components/camera/api21/Camera2SurfaceTextureComponent.kt
@@ -10,10 +10,9 @@ import android.media.ImageReader
 import android.os.Handler
 import android.os.HandlerThread
 import android.util.Log
-import android.view.TextureView
 import androidx.annotation.NonNull
 import androidx.annotation.RequiresApi
-import tv.letsrobot.android.api.components.camera.TextureViewCameraBaseComponent
+import tv.letsrobot.android.api.components.camera.SurfaceTextureCameraBaseComponent
 import tv.letsrobot.android.api.models.CameraSettings
 
 
@@ -21,7 +20,7 @@ import tv.letsrobot.android.api.models.CameraSettings
  * Created by Brendon on 10/6/2018.
  */
 @RequiresApi(21)
-class Camera2TextureComponent(context: Context, settings: CameraSettings, surfaceView: TextureView) : TextureViewCameraBaseComponent(context, settings, surfaceView), ImageReader.OnImageAvailableListener {
+class Camera2SurfaceTextureComponent(context: Context, settings: CameraSettings) : SurfaceTextureCameraBaseComponent(context, settings), ImageReader.OnImageAvailableListener {
     val reader = ImageReader.newInstance(width, height, ImageFormat.JPEG, 10)
 
     private var mPreviewBuilder: CaptureRequest.Builder? = null
@@ -125,12 +124,12 @@ class Camera2TextureComponent(context: Context, settings: CameraSettings, surfac
      * Start the camera preview.
      */
     private fun startPreview() {
-        if (null == mCameraDevice || !textureView.isAvailable/* || null == mPreviewSize*/) {
+        if (null == mCameraDevice) {
             return
         }
         try {
             closePreviewSession()
-            val texture = textureView.surfaceTexture
+            val texture = mStManager.surfaceTexture
             texture.setDefaultBufferSize(height, width)
             mPreviewBuilder = mCameraDevice!!.createCaptureRequest(CameraDevice.TEMPLATE_PREVIEW)
